채팅방 유저 목록 추가

+++ src/main/java/com/icebear2n2/chatroom/user/controller/UserController.java
@@ -103,7 +103,7 @@ public class UserController {
     public String logout(HttpSession session) {
 
         // 세션에서 사용자 ID를 가져와서 uid에 저장
-        Integer uid = (Integer) session.getAttribute("userId");
+        Long uid = (Long) session.getAttribute("userId");
 
         // uid가 존재하면, 세션 무효화
         if (uid != null) {
